Move input measurables into a separate file, and make them extend connection measurables.
