+++ core/renderers/block_rendering_rewrite/block_render_info.js
@@ -34,6 +34,10 @@ goog.require('Blockly.blockRendering.Row');
 goog.require('Blockly.blockRendering.SpacerRow');
 goog.require('Blockly.blockRendering.TopRow');
 
+goog.require('Blockly.blockRendering.InlineInput');
+goog.require('Blockly.blockRendering.ExternalValueInput');
+goog.require('Blockly.blockRendering.StatementInput');
+
 goog.require('Blockly.blockRendering.PreviousConnection');
 goog.require('Blockly.blockRendering.NextConnection');
 goog.require('Blockly.blockRendering.OutputConnection');
